+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changelog (nionui)
 
 0.6.6 (UNRELEASED)
 ------------------
+- Add ability to set top, left, bottom, right, horizontal, and vertical borders independently.
 - Add ability to set top, left, bottom, right, horizontal, and vertical margins independently.
 
 0.6.5 (2022-09-13)